+++ Modules/Microsoft365DSC/DscResources/MSFT_EXOMailboxCalendarFolder/MSFT_EXOMailboxCalendarFolder.psm1
@@ -39,7 +39,7 @@ function Get-TargetResource
         $SharedCalendarSyncStartDate,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',
 
@@ -191,7 +191,7 @@ function Set-TargetResource
         $SharedCalendarSyncStartDate,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',
 
@@ -300,7 +300,7 @@ function Test-TargetResource
         $SharedCalendarSyncStartDate,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',