+++ Modules/Microsoft365DSC/DscResources/MSFT_EXOMailboxFolderPermission/MSFT_EXOMailboxFolderPermission.psm1
@@ -16,7 +16,7 @@ function Get-TargetResource
         $UserPermissions,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',
 
@@ -142,7 +142,7 @@ function Set-TargetResource
         $UserPermissions,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',
 
@@ -267,7 +267,7 @@ function Test-TargetResource
         $UserPermissions,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',