+++ Modules/Microsoft365DSC/DscResources/MSFT_EXOMailboxPlan/MSFT_EXOMailboxPlan.psm1
@@ -48,7 +48,7 @@ function Get-TargetResource
         $RoleAssignmentPolicy,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',
 
@@ -225,7 +225,7 @@ function Set-TargetResource
         $RoleAssignmentPolicy,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',
 
@@ -348,7 +348,7 @@ function Test-TargetResource
         $RoleAssignmentPolicy,
 
         [Parameter()]
-        [ValidateSet('Present')]
+        [ValidateSet('Present','Absent')]
         [System.String]
         $Ensure = 'Present',